requirements.txt
Flask==2.0.1
Flask-Script==2.0.6
python-dotenv
flasgger==0.9.5
pytest==6.2.4
PyJWT==2.3.0
requests==2.27.1